moved code and went trough catalog controller logic
